doc-gen-backend/requirements.txt
amqp==5.3.1
appscript==1.4.0
asgiref==3.11.0
async-timeout==5.0.1
billiard==4.2.3
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-filter==25.2
django-import-export==3.3.7
django-jazzmin==2.6.1
django-redis==6.0.0
django-rest-swagger==2.2.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
docx2pdf==0.1.8
drf-yasg==1.21.11
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
kombu==5.5.4
lxml==6.0.2
MarkupPy==1.18
MarkupSafe==3.0.3
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.5
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
PyJWT==2.10.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
redis==7.1.0
reportlab==4.4.5
requests==2.32.5
simplejson==3.20.2
six==1.17.0
sqlparse==0.5.3
tablib==3.5.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
urllib3==2.5.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
xlrd==2.0.2
xlwt==1.3.0
